requirements.txt
django-nested-inline~=0.4.5
django~=4.1.6
